+++ restrict-imports-gradle-plugin/src/main/java/de/skuzzle/restrictimports/gradle/RestrictedImportsFoundException.java
@@ -1,0 +1,8 @@
+package de.skuzzle.restrictimports.gradle;
+
+public class RestrictedImportsFoundException extends RuntimeException {
+
+    public RestrictedImportsFoundException(String message) {
+        super(message);
+    }
+}