+++ RELEASE_NOTES.md
@@ -1,15 +1,44 @@
-[![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=2.4.0&color=blue)](https://search.maven.org/artifact/de.skuzzle.enforcer/restrict-imports-enforcer-rule/2.4.0/jar)
+[![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=2.5.0-rc-1&color=blue)](https://search.maven.org/artifact/de.skuzzle.enforcer/restrict-imports-enforcer-rule/2.5.0-rc-1/jar)
 
-### Bug fixes
-* [#55](https://github.com/skuzzle/restrict-imports-enforcer-rule/issues/55) Make parallel analysis the default
-* [#63](https://github.com/skuzzle/restrict-imports-enforcer-rule/issues/63) Support for ignoring classes that cannot be fixed
+### Features
+* [#59](https://github.com/skuzzle/restrict-imports-enforcer-rule/issues/59) Provide a Gradle plugin
 
 Maven Central coordinates for this release:
 
+maven
 ```xml
 <dependency>
     <groupId>de.skuzzle.enforcer</groupId>
     <artifactId>restrict-imports-enforcer-rule</artifactId>
-    <version>2.4.0</version>
+    <version>2.5.0-rc-1</version>
 </dependency>
+```
+
+Gradle plugin DSL
+```groovy
+plugins {
+  id("de.skuzzle.restrict.imports") version "2.5.0-rc-1"
+}
+```
+
+Gradle Legacy
+```groovy
+buildscript {
+  repositories {
+    maven {
+      url = uri("https://plugins.gradle.org/m2/")
+    }
+  }
+  dependencies {
+    classpath("de.skuzzle.enforcer:restrict-imports-gradle-plugin:2.5.0-rc-1")
+  }
+}
+
+apply(plugin = "de.skuzzle.restrict.imports")
+```
+
+Gradle version catalog (Toml)
+```toml
+[plugins]
+restrict-imports = { id = "de.skuzzle.restrict.imports", version = "2.5.0-rc-1" }
 ```